Revert "Change prompt git info to rely only on git"
This reverts commit a84d85f4e060348d15fa261f5e213d3b316f92cf.

There are too many issues with this approach:

- Repeated calls to git slows down the prompt (way more than what
  "savings" I ostensibly get by not relying on cut, sed, etc!)
- Untracked flag misbehaves when changing into a child directory
- Prompt breaks right after git init but before committing, since HEAD
  is not a valid ref yet

The previous version is ugly and requires fixing, but relying on Git
plumbing calls is not the way to go.

The git status --porcelain output is already designed to be
script-friendly, so no harm in choosing that approach; the problem is
only in writing elegant code to parse it.