Make vim with fzf robust
The old way of handling editors with fzf relied on an unquoted command
substitution, since this was the only way of allowing for multiple
files. However, filenames with spaces cannot be handled properly with
this method.

To handle this, vz has been promoted from a mere alias to a full
function: it reads in data from fzf --multi --print0 into an array,
since NUL-delimited data is a safe way of working with lists of UNIX
filenames. It then simply invokes vim with the array of files.