Support local completion sources in completion.zsh
I've gotten a lot of mileage out of custom completion sources for some
of my personal scripts. $HOME/.local/share seems as good a path as any
for this, so I added it to my fpath.

None of my completion sources are being checked in, since all of them
are for scripts not included in this repo.