Swap colors 8 and 15 in highlight.zsh
In accordance with a shift across my different configs, I've swapped my
colors 8 and 15 such that color 8 will now be a foreground gray, since
some non-theme-able programs assume it that way.

Color 15 is still problematic since I use it as my background gray, but
non-theme-able programs assume it to be bright white. A more solid
solution involving a return to Xterm's standard 256-color palette might
be in order soon.