Rework git branch parsing in highlight.zsh
Apparently, the former match-based parsing doesn't work on zsh from
Homebrew (don't ask...), so the parsing now relies on a series of string
manipulations.

I actually like it better this way. For the branch parsing at the very
least, trimming with parameter expansions is so much simpler than a
capturing regex.

For ahead/behind status, ehh... The parsing loop is elegant and general,
but seems pretty overkill for the very restrictive input space ("ahead
n", "behind n", or "ahead m, behind n"). It's late and I'm tired; if an
epiphany comes to me in the morning, I'll change it, otherwise I'm
keeping it.